Implemented a module hierarchy
