+++ controller/__init__.py
@@ -1,0 +1,3 @@
+from .control import *
+
+__all__ = ["controller"]
