+++ model/__init__.py
@@ -1,0 +1,3 @@
+from .model import *
+
+__all__ = ['model']
