+++ slayTheHero.py
@@ -1,0 +1,4 @@
+from controller import *
+
+if __name__ == '__main__':
+    c = controller()
