+++ view/__init__.py
@@ -1,0 +1,4 @@
+from .menus import *
+from .view import *
+
+__all__ = ['window', 'menu', 'title', 'button', 'menuList', 'menuFont']
